# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/admin/board": {
            "get": {
                "tags": [
                    "게시판"
                ],
                "summary": "게시판 목록",
                "description": "게시판 목록",
                "operationId": "adminBoardList",
                "responses": {
                    "200": {
                        "description": "successfully Modified",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BoardOptionJson"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "게시판"
                ],
                "summary": "게시판 생성",
                "description": "게시판 생성",
                "operationId": "adminBoardCreate",
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type"
                                ],
                                "properties": {
                                    "name": {
                                        "$ref": "#/components/schemas/Board/properties/name"
                                    },
                                    "type": {
                                        "$ref": "#/components/schemas/Board/properties/type"
                                    },
                                    "hidden": {
                                        "$ref": "#/components/schemas/Board/properties/hidden"
                                    },
                                    "options": {
                                        "$ref": "#/components/schemas/BoardOptionJson/properties/options"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successfully Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "successfully Modified"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "failed registered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "statusCode": {
                                                    "properties": {
                                                        "20010": {
                                                            "description": "이미 존재하는 type 입니다.",
                                                            "properties": {
                                                                "key": {
                                                                    "description": "type",
                                                                    "type": "string",
                                                                    "example": "type"
                                                                },
                                                                "message": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "type": "object"
                                                        },
                                                        "20401": {
                                                            "description": "옵션 및 옵션 값을 확인해주세요.",
                                                            "properties": {
                                                                "key": {
                                                                    "description": "options.editor.value",
                                                                    "type": "string",
                                                                    "example": "options.editor.value"
                                                                },
                                                                "message": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/board/{id}": {
            "patch": {
                "tags": [
                    "게시판"
                ],
                "summary": "게시판 수정",
                "description": "게시판 수정",
                "operationId": "adminBoardModify",
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "$ref": "#/components/schemas/Board/properties/name"
                                    },
                                    "hidden": {
                                        "$ref": "#/components/schemas/Board/properties/hidden"
                                    },
                                    "options": {
                                        "$ref": "#/components/schemas/BoardOptionJson/properties/options"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successfully Modified",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "successfully Modified"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "failed registered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "statusCode": {
                                                    "properties": {
                                                        "20401": {
                                                            "description": "옵션 및 옵션 값을 확인해주세요.",
                                                            "properties": {
                                                                "key": {
                                                                    "description": "options.editor.value",
                                                                    "type": "string",
                                                                    "example": "options.editor.value"
                                                                },
                                                                "message": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BoardOptionJson": {
                "properties": {
                    "id": {
                        "description": "게시판 고유번호",
                        "type": "string",
                        "example": 7
                    },
                    "name": {
                        "description": "게시판 명",
                        "type": "string",
                        "example": "공지사항"
                    },
                    "type": {
                        "description": "게시판 타입",
                        "type": "string",
                        "example": "notice"
                    },
                    "options": {
                        "properties": {
                            "board": {
                                "description": "작성 권한<br/>all:모두 작성가능<br/>manager:운영진만 작성가능<br/>member:회원만 작성가능",
                                "type": "enum",
                                "example": "all"
                            },
                            "boardReply": {
                                "description": "글 답변 작성<br/>0:사용안함<br/>1:사용함",
                                "type": "enum",
                                "example": "0"
                            },
                            "secret": {
                                "description": "시크릿 사용<br/>0:사용안함<br/>1:사용함",
                                "type": "enum",
                                "example": "0"
                            },
                            "listViewContent": {
                                "description": "목록에서 내용보기<br/>0:사용안함<br/>1:사용함",
                                "type": "enum",
                                "example": "0"
                            },
                            "reply": {
                                "description": "댓글 사용<br/>0:사용안함<br/>1:사용함",
                                "type": "enum",
                                "example": "1"
                            },
                            "editor": {
                                "description": "에디터<br/>all:모두 사용<br/>ck:CK에디터 사용<br/>markd:마크다운 사용",
                                "type": "enum",
                                "example": "all"
                            },
                            "image": {
                                "description": "이미지 첨부<br/>0:사용안함<br/>1:사용함",
                                "type": "enum",
                                "example": "0"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "게시판 옵션"
                }
            }
        },
        "securitySchemes": {
            "davinci_auth": {
                "type": "http",
                "scheme": "bearer"
            },
            "sample_auth": {
                "type": "oauth2",
                "flows": {
                    "implicit": {
                        "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
                        "scopes": {
                            "read:pets": "read your pets",
                            "write:pets": "modify pets in your account"
                        }
                    }
                }
            }
        }
    },
...
    "security": [
        []
    ]
}